commit additional changes for initial menu
