+++ src/LottoFun/view/Header.java
@@ -1,0 +1,25 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package LottoFun.view;
+
+/**
+ *
+ * @author t.bain
+ */
+public class Header {
+    public Header()
+    {
+        
+    }
+    
+    public void displayHeader(String screenName)
+    {
+        System.out.println("Lotto Fun \t\t\t" + screenName);
+        System.out.println("");
+        System.out.println("");
+        
+    }
+}
